+++ libs/zIndex.ts
@@ -1,0 +1,4 @@
+export const rootZIndex = {
+  zIndexHeader: 1,
+  zIndexModal: 2
+};
