+++ src/data/pageList.json
@@ -1,0 +1,22 @@
+[
+  {
+    "title": "works",
+    "link": "/works"
+  },
+  {
+    "title": "huga",
+    "link": "/a"
+  },
+  {
+    "title": "piyo",
+    "link": "/b"
+  },
+  {
+    "title": "aiueo",
+    "link": "/c"
+  },
+  {
+    "title": "aaaaaa",
+    "link": "/d"
+  }
+]
